+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/PodcastManagerImpl.kt
@@ -489,6 +489,14 @@ class PodcastManagerImpl @Inject constructor(
     override suspend fun updateAutoDownload(podcastUuids: Collection<String>, isEnabled: Boolean) {
         val status = if (isEnabled) Podcast.AUTO_DOWNLOAD_NEW_EPISODES else Podcast.AUTO_DOWNLOAD_OFF
         podcastDao.updateAutoDownloadStatus(podcastUuids, status)
+        if (isEnabled) {
+            settings.autoDownloadNewEpisodes.set(Podcast.AUTO_DOWNLOAD_NEW_EPISODES, updateModifiedAt = false)
+        } else {
+            val podcasts = podcastDao.findSubscribedNoOrder()
+            if (podcasts.none { it.isAutoDownloadNewEpisodes }) {
+                settings.autoDownloadNewEpisodes.set(Podcast.AUTO_DOWNLOAD_OFF, updateModifiedAt = false)
+            }
+        }
     }
 
     override suspend fun updateAllShowNotifications(showNotifications: Boolean) {
@@ -500,6 +508,14 @@ class PodcastManagerImpl @Inject constructor(
 
     override fun updateAutoDownloadStatusBlocking(podcast: Podcast, autoDownloadStatus: Int) {
         podcastDao.updateAutoDownloadStatusBlocking(autoDownloadStatus, podcast.uuid)
+        if (autoDownloadStatus == Podcast.AUTO_DOWNLOAD_NEW_EPISODES) {
+            settings.autoDownloadNewEpisodes.set(autoDownloadStatus, updateModifiedAt = false)
+        } else {
+            val podcasts = podcastDao.findSubscribedBlocking()
+            if (podcasts.none { it.isAutoDownloadNewEpisodes }) {
+                settings.autoDownloadNewEpisodes.set(autoDownloadStatus, updateModifiedAt = false)
+            }
+        }
     }
 
     override suspend fun updateAutoAddToUpNext(podcast: Podcast, autoAddToUpNext: Podcast.AutoAddUpNext) {
@@ -577,6 +593,10 @@ class PodcastManagerImpl @Inject constructor(
     }
 
     override fun checkForEpisodesToDownloadBlocking(episodeUuidsAdded: List<String>, downloadManager: DownloadManager) {
+        if (settings.autoDownloadNewEpisodes.value == Podcast.AUTO_DOWNLOAD_OFF) {
+            return
+        }
+
         Timber.i("Auto download podcasts checkForEpisodesToDownload. Episodes %s", episodeUuidsAdded.size)
 
         val podcastUuidToAutoDownload = HashMap<String, Boolean>()