+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 8.7
 -----
-
+*   Bug Fixes
+    *   Respect global auto download setting
+        ([#4957](https://github.com/Automattic/pocket-casts-android/pull/4957))
 
 8.6
 -----